requirements-airflow.txt
pandas>=2.0.0
requests>=2.31.0
# sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
yfinance>=0.2.28
beautifulsoup4>=4.12.0

